LayoutTests/storage/indexeddb/modern/deleteindex-1.html

This tests deleting an index and then committing the transaction.<br>
<div id="logger"></div>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    log("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

function log(message)
{
    document.getElementById("logger").innerHTML += message + "<br>";
}

var index;
var objectStore;
var database;

function checkIndexValues()
{
    var countRequest = index.count();
    countRequest.onsuccess = function() {
        log("Count is: " + countRequest.result);
    }

    var cursorRequest = index.openCursor();
    cursorRequest.onsuccess = function() {
        var cursor = cursorRequest.result;
        log("Cursor at record: " + cursor.key + " / " + cursor.primaryKey);
        
        if (cursor.key != undefined)
            cursor.continue();
        else {
            objectStore.deleteIndex("TestIndex1");
            log("Deleted the index");
        }
    }
    cursorRequest.onerror = function(e) {
        log("Unexpected error opening or iterating cursor");
        logCursor(cursorRequest.result);
        done();
    } 
}

var createRequest = window.indexedDB.open("DeleteIndex1Database", 1);
createRequest.onupgradeneeded = function(event) {
    log("Initial upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

    var versionTransaction = createRequest.transaction;
    database = event.target.result;
    objectStore = database.createObjectStore("TestObjectStore");
    objectStore.put({ bar: "A" }, 1);
    objectStore.put({ bar: "A" }, 2);

    index = objectStore.createIndex("TestIndex1", "bar");
    
    checkIndexValues();
    
    versionTransaction.onabort = function(event) {
        log("Initial upgrade versionchange transaction unexpected abort");
        done();
    }

    versionTransaction.oncomplete = function(event) {
        log("Initial upgrade versionchange transaction complete");
        continueTest();
    }

    versionTransaction.onerror = function(event) {
        log("Initial upgrade versionchange transaction unexpected error" + event);
        done();
    }
}

function continueTest() {
    var transaction = database.transaction("TestObjectStore");
    objectStore = transaction.objectStore("TestObjectStore");

    var names = objectStore.indexNames;
    log("Object store has indexes:")
    for (var i = 0; i < names.length; ++i)
        log(names[i]);
    
    try {
        objectStore.index("TestIndex1");
    } catch(e) {
        log("Unable to get index from object store (it shouldn't exist)");
    }

    transaction.onabort = function(event) {
        log("Transaction unexpected abort");
        done();
    }

    transaction.oncomplete = function(event) {
        log("Transaction complete");
        done();
    }

    transaction.onerror = function(event) {
        log("Transaction unexpected error" + event);
        done();
    }
}

</script>
