LayoutTests/storage/indexeddb/modern/deleteindex-1-expected.txt

This tests deleting an index and then committing the transaction.
Initial upgrade needed: Old version - 0 New version - 1
Count is: 2
Cursor at record: A / 1
Cursor at record: A / 2
Cursor at record: undefined / undefined
Deleted the index
Initial upgrade versionchange transaction complete
Object store has indexes:
Unable to get index from object store (it shouldn't exist)
Transaction complete
Done
